Feat : 일정에 유저 연관관계 매핑

+++ src/main/java/sparta/advancedscheduler/schedule/service/ScheduleService.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import sparta.advancedscheduler.auth.service.AuthorizationService;
 import sparta.advancedscheduler.comment.service.CommentService;
 import sparta.advancedscheduler.schedule.dto.RequestScheduleDto;
 import sparta.advancedscheduler.schedule.dto.RequestScheduleUpdateDto;
@@ -14,6 +15,8 @@ import sparta.advancedscheduler.schedule.dto.ResponseScheduleDto;
 import sparta.advancedscheduler.schedule.dto.ResponseScheduleListDto;
 import sparta.advancedscheduler.schedule.entity.Schedule;
 import sparta.advancedscheduler.schedule.repository.ScheduleRepository;
+import sparta.advancedscheduler.user.entity.User;
+import sparta.advancedscheduler.user.service.UserService;
 
 import java.util.List;
 
@@ -23,15 +26,20 @@ import java.util.List;
 public class ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
+    private final AuthorizationService authorizationService;
     private final ScheduleFinder scheduleFinder;
     private final CommentService commentService;
+    private final UserService userService;
+
 
     @Transactional
-    public Long createSchedule(RequestScheduleDto requestScheduleDto) {
+    public Long createSchedule(Long userId , RequestScheduleDto requestScheduleDto) {
+        User user = userService.getUserById(userId);
+
         Schedule schedule = new Schedule(
                 requestScheduleDto.getTitle(),
                 requestScheduleDto.getContent(),
-                requestScheduleDto.getPoster()
+                user
         );
         scheduleRepository.save(schedule);
 
@@ -40,14 +48,14 @@ public class ScheduleService {
 
 
     @Transactional(readOnly = true)
-    public Page<ResponseScheduleListDto> findScheduleListByName(String poster , Pageable pageable) {
-        Page<Schedule> schedules = scheduleRepository.findAllByPosterOrderByLastModifiedDateDesc(poster,pageable);
+    public Page<ResponseScheduleListDto> findScheduleListByName(Long userID , Pageable pageable) {
+        Page<Schedule> schedules = scheduleRepository.findAllByUserIdOrderByLastModifiedDateDesc(userID,pageable);
 
         List<ResponseScheduleListDto> dtoList = schedules.getContent().stream()
                 .map(schedule -> new ResponseScheduleListDto(
                             schedule.getTitle(),
                             schedule.getContent(),
-                            schedule.getPoster(),
+                            schedule.getUser().getUsername(),
                             commentService.getCommentCount(schedule.getId()),
                             schedule.getCreatedDate(),
                             schedule.getLastModifiedDate()
@@ -69,29 +77,30 @@ public class ScheduleService {
         return new ResponseScheduleDto(
                 schedule.getTitle(),
                 schedule.getContent(),
-                schedule.getPoster(),
+                schedule.getUser().getUsername(),
                 schedule.getCreatedDate(),
                 schedule.getLastModifiedDate()
         );
     }
 
     @Transactional
-    public ResponseScheduleDto updateScheduleById(Long scheduleId , RequestScheduleUpdateDto body) {
-
+    public ResponseScheduleDto updateScheduleById(Long userId , Long scheduleId , RequestScheduleUpdateDto body) {
         Schedule schedule = scheduleFinder.findScheduleById(scheduleId);
+        authorizationService.checkAuthorization(schedule.getUser().getId() ,userId);
 
         schedule.update(body);
         return new ResponseScheduleDto(
                 schedule.getTitle(),
                 schedule.getContent(),
-                schedule.getPoster(),
+                schedule.getUser().getUsername(),
                 schedule.getCreatedDate(),
                 schedule.getLastModifiedDate()
         );
     }
     @Transactional
-    public Long deleteSchedule(Long scheduleId) {
+    public Long deleteSchedule(Long userId,Long scheduleId) {
         Schedule schedule = scheduleFinder.findScheduleById(scheduleId);
+        authorizationService.checkAuthorization(schedule.getUser().getId() ,userId);
         scheduleRepository.delete(schedule);
         return scheduleId;
     }
